requirements.txt
chromadb
sentence-transformers
huggingface_hub
python-dotenv
numpy